pyproject.toml
[tool.poetry]
name = "vmc"
version = "0.0.0"
description = "VMC(Virtual Model Collection) is a collection of virtual models for various tasks."
authors = ["HSPK <whxway@whu.edu.cn>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "vmc" }]

[tool.poetry.scripts]
vmc = 'vmc.cli:cli'

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
zhipuai = "^2.1.3"
openai = "^1.35.14"
fastapi = { extras = ["all"], version = "^0.111.1" }
uvicorn = { extras = ["standard"], version = "^0.30.1" }
gunicorn = "^22.0.0"
pymongo = "^4.8.0"
pydantic = "^2.0.0"
google-generativeai = "^0.7.2"
loguru = "^0.7.2"
tiktoken = "^0.7.0"
arrow = "^1.3.0"
slark = "^0.1.4"

[tool.poetry.extras]
local = ["transformers", "torch"]

[tool.poetry.group.dev.dependencies]
pdoc = "^14.5.1"
pre-commit = "^3.8.0"
ruff = "^0.6.9"


[tool.poetry.group.test.dependencies]
ipython = "^8.26.0"
jupyter = "^1.0.0"
pytest = "^8.2.2"



[tool.poetry.group.serve.dependencies]
transformers = "^4.46.0"
torch = "^2.5.0"
sentence-transformers = "^3.2.1"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
